refactor: remove AdminOrdersPage and implement order detail components
- Deleted AdminOrdersPage component and its associated logic.
- Added OrderDetailHeader, OrderDetailItems, OrderDetailPayment, OrderDetailSummary, and OrderDetailTotal components to handle order details.
- Improved status handling and UI for order actions (confirm, ship, cancel).
- Enhanced payment details display with proof image functionality.
- Updated styling for better user experience and consistency.
